MNES-1203 save all at once
