+++ src/app/[locale]/settings/_components/id-settings/IdSettingsCard.tsx
@@ -3,19 +3,27 @@ import { alpha, Box, Skeleton, Divider, Typography, styled, Button } from '@mui/
 import { CardHeading } from 'components/basics/CardHeading';
 import { messages } from 'lib/i18n/localization';
 import { Fragment, useEffect, useState } from 'react';
-import { FormattedMessage } from 'react-intl';
+import { FormattedMessage, useIntl } from 'react-intl';
 import { IdGenerationSettingFrontend } from 'lib/types/IdGenerationSettingFrontend';
 import { IdSettingEntry } from './IdSettingEntry';
 import { AssetIdRedirectDocumentationDialog } from './AssetIdRedirectDocumentationDialog';
 import CloseIcon from '@mui/icons-material/Close';
 import CheckIcon from '@mui/icons-material/Check';
 import EditIcon from '@mui/icons-material/Edit';
-import { SubmitHandler, useFieldArray, useForm } from 'react-hook-form';
-import { ConnectionFormData } from 'app/[locale]/settings/_components/mnestix-connections/MnestixConnectionsCard';
+import { useFieldArray, useForm } from 'react-hook-form';
 import { useAsyncEffect } from 'lib/hooks/UseAsyncEffect';
-import { isValidIdPrefix, isValidShortIdPrefix } from 'lib/util/IdValidationUtil';
+import { showError } from 'lib/util/ErrorHandlerUtil';
+import {
+    ISubmodelElement, Property,
+    Qualifier,
+    SubmodelElementCollection
+} from '@aas-core-works/aas-core3.0-typescript/dist/types/types';
+import { getArrayFromString } from 'lib/util/SubmodelResolverUtil';
+import { useAuth } from 'lib/hooks/UseAuth';
+import { useApis } from 'components/azureAuthentication/ApiProvider';
+import { useNotificationSpawner } from 'lib/hooks/UseNotificationSpawner';
 
-const StyledDocumentationButton = styled(Box)(({ theme }) => ({
+const StyledDocumentationButton = styled(Box)(({theme}) => ({
     display: 'flex',
     alignItems: 'center',
     backgroundColor: alpha(theme.palette.primary.main, 0.1),
@@ -32,45 +40,117 @@ const StyledDocumentationButton = styled(Box)(({ theme }) => ({
     },
 }));
 
-type IdSettingsCardProps = {
-    readonly idSettings: IdGenerationSettingFrontend[] | undefined;
-    readonly isLoading?: boolean;
-    readonly handleChange: (idShort: string, values: { prefix: string; dynamicPart: string }) => Promise<void>;
-};
-
 export type IdSettingsFormData = {
     idSettings: IdGenerationSettingFrontend[];
 }
 
-export function IdSettingsCard(props: IdSettingsCardProps) {
-    const settings = props.idSettings || [];
+export function IdSettingsCard() {
     const [isEditMode, setIsEditMode] = useState(false);
     const [documentationModalOpen, setDocumentationModalOpen] = useState(false);
+    const auth = useAuth();
+    const bearerToken = auth.getBearerToken();
+    const {configurationClient} = useApis();
+    const notificationSpawner = useNotificationSpawner();
+    const intl = useIntl();
+    const [isLoading, setIsLoading] = useState(false);
+    const [settings, setSettings] = useState<IdGenerationSettingFrontend[]>([]);
 
     const {
         register,
         control,
         handleSubmit,
         reset,
-        formState: { errors },
-    } = useForm<IdSettingsFormData>({defaultValues: {idSettings: settings} });
-    
-    useEffect(() => {
-        reset({idSettings: settings})
-    }, [props.isLoading])
+        formState: {errors},
+    } = useForm<IdSettingsFormData>({defaultValues: {idSettings: settings}});
 
-    const { fields } = useFieldArray<IdSettingsFormData>({
+    const {fields} = useFieldArray<IdSettingsFormData>({
         control,
         name: 'idSettings',
     });
     
-    async function saveIdSettings(data: IdSettingsFormData) {
-        for (const setting of data.idSettings) {
-            if(setting.prefix.value && setting.dynamicPart.value) {
-                await props.handleChange(setting.name, {prefix: setting.prefix.value, dynamicPart: setting.dynamicPart.value})
-            }
+    // Fetch id settings initially
+    useAsyncEffect(async () => {
+        await fetchIdSettings();
+    }, [bearerToken]);
+
+    const fetchIdSettings = async () => {
+        try {
+            setIsLoading(true);
+            const res = await configurationClient.getIdGenerationSettings();
+            const _settings: IdGenerationSettingFrontend[] = [];
+            // set settings from api response
+            res.submodelElements?.forEach((el) => {
+                const element = el as ISubmodelElement;
+                const collection = el as SubmodelElementCollection;
+                const _settingsList = collection.value;
+                const name = el.idShort;
+
+                // IdType (to apply correct validation)
+                const idTypeQualifier = element.qualifiers?.find((q: Qualifier) => {
+                    return q.type === 'SMT/IdType';
+                });
+                const idType = idTypeQualifier?.value;
+
+                const prefix = _settingsList?.find((e) => e.idShort === 'Prefix') as Property;
+                const dynamicPart = _settingsList?.find((e) => e.idShort === 'DynamicPart') as Property;
+
+                const dynamicPartAllowedQualifier = dynamicPart?.qualifiers?.find((q: Qualifier) => {
+                    return q.type === 'AllowedValue';
+                });
+                const allowedDynamicPartValues = getArrayFromString(dynamicPartAllowedQualifier?.value || '');
+
+                const prefixExampleValueQualifier = prefix?.qualifiers?.find((q: Qualifier) => {
+                    return q.type === 'ExampleValue';
+                });
+                const prefixExampleValue = prefixExampleValueQualifier?.value;
+
+                _settings.push({
+                    name: name || '',
+                    idType,
+                    prefix: {
+                        value: prefix?.value,
+                        exampleValue: prefixExampleValue,
+                    },
+                    dynamicPart: {
+                        value: dynamicPart?.value,
+                        allowedValues: allowedDynamicPartValues,
+                        // (we do not fill example value from api currently)
+                    },
+                });
+            });
+            setSettings(_settings);
+            // set form state
+            reset({idSettings: _settings})
+
+        } catch (e) {
+            showError(e, notificationSpawner);
+        } finally {
+            setIsLoading(false);
         }
-        setIsEditMode(false); // todo move save/load to card and extract form in extra component
+    };
+    
+    async function saveIdSettings(data: IdSettingsFormData) {
+        try {
+            setIsLoading(true);
+            for (const setting of data.idSettings) {
+                if (setting.prefix.value && setting.dynamicPart.value) {
+                    await configurationClient.putSingleIdGenerationSetting(setting.name, bearerToken, {
+                        prefix: setting.prefix.value,
+                        dynamicPart: setting.dynamicPart.value
+                    })
+                }
+            }
+            await fetchIdSettings();
+            notificationSpawner.spawn({
+                message: intl.formatMessage(messages.mnestix.successfullyUpdated),
+                severity: 'success',
+            });
+            setIsEditMode(false);
+        } catch (e) {
+            showError(e, notificationSpawner);
+        } finally {
+            setIsLoading(false);
+        }
     }
 
     const cancelEdit = () => {
@@ -79,7 +159,7 @@ export function IdSettingsCard(props: IdSettingsCardProps) {
     };
 
     return (
-        <Box sx={{ p: 3, width: '100%' }}>
+        <Box sx={{p: 3, width: '100%'}}>
             <Box display="flex" flexDirection="row" justifyContent="space-between">
                 <CardHeading
                     title={<FormattedMessage {...messages.mnestix.idStructure} />}
@@ -88,33 +168,35 @@ export function IdSettingsCard(props: IdSettingsCardProps) {
                 <Box display="flex" gap={2} alignContent="center" flexWrap="wrap">
                     {isEditMode ? (
                         <>
-                            <Button variant="outlined" startIcon={<CloseIcon />} onClick={()=> {cancelEdit()}}>
+                            <Button variant="outlined" startIcon={<CloseIcon/>} onClick={() => {
+                                cancelEdit()
+                            }}>
                                 <FormattedMessage {...messages.mnestix.cancel} />
                             </Button>
                             <Button
                                 variant="contained"
-                                startIcon={<CheckIcon />}
+                                startIcon={<CheckIcon/>}
                                 onClick={handleSubmit((data) => saveIdSettings(data))}
                             >
                                 <FormattedMessage {...messages.mnestix.connections.saveButton} />
                             </Button>
                         </>
                     ) : (
-                        <Button variant="contained" startIcon={<EditIcon />} onClick={() => setIsEditMode(true)}>
+                        <Button variant="contained" startIcon={<EditIcon/>} onClick={() => setIsEditMode(true)}>
                             <FormattedMessage {...messages.mnestix.connections.editButton} />
                         </Button>
                     )}
                 </Box>
             </Box>
-            <Box sx={{ my: 2 }}>
-                <Divider />
-                {props.isLoading &&
+            <Box sx={{my: 2}}>
+                <Divider/>
+                {isLoading &&
                     !settings.length &&
                     [0, 1, 2, 3, 4].map((i) => {
                         return (
                             <Fragment key={i}>
-                                <Skeleton variant="text" width="50%" height={26} sx={{ m: 2 }} />
-                                {i < 4 && <Divider />}
+                                <Skeleton variant="text" width="50%" height={26} sx={{m: 2}}/>
+                                {i < 4 && <Divider/>}
                             </Fragment>
                         );
                     })}
@@ -124,9 +206,8 @@ export function IdSettingsCard(props: IdSettingsCardProps) {
                             key={index}
                             index={index}
                             field={field}
-                            handleChange={props.handleChange}
                             editMode={isEditMode}
-                            isLoading={props.isLoading}
+                            isLoading={isLoading}
                             control={control}
                             errors={errors}
                             register={register}
@@ -134,9 +215,9 @@ export function IdSettingsCard(props: IdSettingsCardProps) {
                     );
                 })}
             </Box>
-            <Box sx={{ display: 'flex' }}>
+            <Box sx={{display: 'flex'}}>
                 <StyledDocumentationButton onClick={() => setDocumentationModalOpen(true)}>
-                    <InfoOutlined sx={{ mr: 1 }} />
+                    <InfoOutlined sx={{mr: 1}}/>
                     <Typography>
                         <FormattedMessage {...messages.mnestix.assetIdDocumentation.title} />
                     </Typography>
